Move database and converters under repository package
