+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/repository/builder/RepositoryFactory.kt
@@ -1,8 +1,22 @@
-package io.getstream.chat.android.livedata.repository
+package io.getstream.chat.android.livedata.repository.builder
 
 import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.client.models.User
-import io.getstream.chat.android.livedata.ChatDatabase
+import io.getstream.chat.android.livedata.repository.ChannelConfigRepository
+import io.getstream.chat.android.livedata.repository.ChannelConfigRepositoryImpl
+import io.getstream.chat.android.livedata.repository.ChannelRepository
+import io.getstream.chat.android.livedata.repository.ChannelRepositoryImpl
+import io.getstream.chat.android.livedata.repository.MessageRepository
+import io.getstream.chat.android.livedata.repository.MessageRepositoryImpl
+import io.getstream.chat.android.livedata.repository.QueryChannelsRepository
+import io.getstream.chat.android.livedata.repository.QueryChannelsRepositoryImpl
+import io.getstream.chat.android.livedata.repository.ReactionRepository
+import io.getstream.chat.android.livedata.repository.ReactionRepositoryImpl
+import io.getstream.chat.android.livedata.repository.SyncStateRepository
+import io.getstream.chat.android.livedata.repository.SyncStateRepositoryImpl
+import io.getstream.chat.android.livedata.repository.UserRepository
+import io.getstream.chat.android.livedata.repository.UserRepositoryImpl
+import io.getstream.chat.android.livedata.repository.database.ChatDatabase
 
 internal class RepositoryFactory(
     private val database: ChatDatabase,
