+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/repository/database/converter/MapConverter.kt
@@ -1,4 +1,4 @@
-package io.getstream.chat.android.livedata.converter
+package io.getstream.chat.android.livedata.repository.database.converter
 
 import androidx.room.TypeConverter
 import com.google.gson.reflect.TypeToken
