+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/repository/database/converter/SyncStatusConverter.kt
@@ -1,4 +1,4 @@
-package io.getstream.chat.android.livedata.converter
+package io.getstream.chat.android.livedata.repository.database.converter
 
 import androidx.room.TypeConverter
 import io.getstream.chat.android.client.utils.SyncStatus
